feat: automatically generate Django secret
The Django secret should *always* be a strong secret value, and because
the user should never need to enter it directly, we should automatically
generate it for them to ensure it is strong. As a reminder, the Django
secret is never used for login; it is used for cryptographic operations.

This change adds a prompt to ask if the script should automatically
generate the secret, and this option can automatically be selected by
specifying the '-y' flag upon run.

This also fixes at least two ShellCheck notices:

* https://www.shellcheck.net/wiki/SC2162
* https://www.shellcheck.net/wiki/SC2059