fix(?): remove systemd requirement
Removing systemd as a requirement mitigates an edge case where disconected
users could overiding installed systemd and, in the worst case scenario,
bork a RHEL installation.

Even though systemd is technically a requirement, it is safe to assume
any supported RHEL has systemd. This also aligns with packages like
podman, which rely on systemd, but don't have it declared as a
requirement.