Set --pids-limit explicitly
On RHEL 8, you have to set --pids-limit explicitly, or containers fail
to start. This can be done in command line or through one of many
supported `containers.conf` files. However, our installer is the only
place where we tell users this is necessary, and the message is hidden
behind a check for cgroups v2 - but you _can_ have cgroups v2 without
`containers.conf`, and then nothing is stopping you from running the
installer and starting the containers. Except they won't work.