feat: add "check" command
Sanity check Quipucords setup and configurations.
This command will check and print the status of important
files and directories that the installer creates for running Quipucords.

Example of status: OK, Not owned by you and Missing.
The return code for the command is the number of problematic files/dirs.

Relates to JIRA: DISCOVERY-729