Fix env file overrides
The logic in conditional was inverted - we checked if there isn't a line
starting with a variable name, and if there wasn't, we modified that
line. If there was, we added new line.

Parameter pattern substitution takes a shell glob, not regular
expression. I tried using different glob and had it working from
interactive shell, but not in the script. I ported the entire thing to
sed, which is already used elsewhere in the script and works well.